+++ src/Shrun/Notify/MonadDBus.hs
@@ -1,13 +1,17 @@
 -- | Effect for DBus.
 module Shrun.Notify.MonadDBus
   ( MonadDBus (..),
+    notifyDBus,
   )
 where
 
 import DBus.Client (Client)
 import DBus.Client qualified as DBusC
-import DBus.Notify (Note, Notification)
+import DBus.Notify (Hint (Urgency), Note (..), Notification)
 import DBus.Notify qualified as DBusN
+import Data.Text qualified as T
+import Shrun.Notify.MonadNotify (ShrunNote)
+import Shrun.Notify.Types (NotifyTimeout (..))
 import Shrun.Prelude
 
 -- | Effect for DBus.
@@ -25,3 +29,23 @@ instance MonadDBus IO where
 instance (MonadDBus m) => MonadDBus (ReaderT env m) where
   connectSession = lift connectSession
   notify c = lift . notify c
+
+notifyDBus :: (MonadDBus m) => Client -> ShrunNote -> m ()
+notifyDBus client = void . notify client . shrunToDBus
+
+shrunToDBus :: ShrunNote -> Note
+shrunToDBus shrunNote =
+  DBusN.Note
+    { appName = "Shrun",
+      summary = unpack $ shrunNote ^. #summary,
+      body = Just . DBusN.Text . T.unpack $ shrunNote ^. #body,
+      appImage = Nothing,
+      hints = [Urgency (shrunNote ^. #urgency)],
+      expiry,
+      actions = []
+    }
+  where
+    expiry = case shrunNote ^. #timeout of
+      NotifyTimeoutNever -> DBusN.Never
+      NotifyTimeoutSeconds s ->
+        DBusN.Milliseconds $ 1_000 * fromIntegral s