+++ src/Shrun/ShellT.hs
@@ -7,17 +7,16 @@ module Shrun.ShellT
   )
 where
 
-import DBus.Notify (Hint (Urgency), Note (..), UrgencyLevel (..))
-import DBus.Notify qualified as DBusN
-import Data.Text qualified as T
 import Shrun.Configuration.Env.Types (Env)
 import Shrun.Logging.MonadRegionLogger (MonadRegionLogger (..))
+import Shrun.Notify.MonadAppleScript (MonadAppleScript)
+import Shrun.Notify.MonadAppleScript qualified as MonadAppleScript
 import Shrun.Notify.MonadDBus (MonadDBus)
 import Shrun.Notify.MonadDBus qualified as MonadDBus
-import Shrun.Notify.MonadNotify (MonadNotify (..), ShrunNote)
+import Shrun.Notify.MonadNotify (MonadNotify (..))
 import Shrun.Notify.MonadNotifySend (MonadNotifySend)
 import Shrun.Notify.MonadNotifySend qualified as MonadNotifySend
-import Shrun.Notify.Types (NotifySystem (..), NotifyTimeout (..))
+import Shrun.Notify.Types (NotifySystem (..))
 import Shrun.Prelude
 
 -- | `ShellT` is the main application type that runs shell commands.
@@ -27,6 +26,7 @@ newtype ShellT env m a = MkShellT (ReaderT env m a)
     ( Functor,
       Applicative,
       Monad,
+      MonadAppleScript,
       MonadAsync,
       MonadDBus,
       MonadCatch,
@@ -62,51 +62,15 @@ runShellT (MkShellT rdr) = runReaderT rdr
 
 deriving newtype instance (MonadRegionLogger m) => MonadRegionLogger (ShellT Env m)
 
-instance (MonadDBus m, MonadNotifySend m) => MonadNotify (ShellT Env m) where
+instance
+  ( MonadAppleScript m,
+    MonadDBus m,
+    MonadNotifySend m
+  ) =>
+  MonadNotify (ShellT Env m)
+  where
   notify note = asks (preview (#notifyEnv %? #system)) >>= traverse_ sendNote
     where
-      sendNote (DBus client) = void $ MonadDBus.notify client (shrunToDBus note)
-      sendNote NotifySend = MonadNotifySend.notify (shrunToNotifySend note)
-
-shrunToDBus :: ShrunNote -> Note
-shrunToDBus shrunNote =
-  DBusN.Note
-    { appName = "Shrun",
-      summary = unpack $ shrunNote ^. #summary,
-      body = Just . DBusN.Text . T.unpack $ shrunNote ^. #body,
-      appImage = Nothing,
-      hints = [Urgency (shrunNote ^. #urgency)],
-      expiry,
-      actions = []
-    }
-  where
-    expiry = case shrunNote ^. #timeout of
-      NotifyTimeoutNever -> DBusN.Never
-      NotifyTimeoutSeconds s ->
-        DBusN.Milliseconds $ 1_000 * fromIntegral s
-
-shrunToNotifySend :: ShrunNote -> Text
-shrunToNotifySend shrunNote = txt
-  where
-    txt =
-      mconcat
-        [ "notify-send ",
-          " --app-name Shrun \"",
-          shrunNote ^. #summary,
-          "\" ",
-          (\b -> " \"" <> b <> "\" ") (shrunNote ^. #body),
-          ulToNS (shrunNote ^. #urgency),
-          timeout
-        ]
-
-    ulToNS Low = " --urgency low "
-    ulToNS Normal = " --urgency normal "
-    ulToNS Critical = " --urgency critical "
-
-    timeout = case shrunNote ^. #timeout of
-      NotifyTimeoutNever -> " --expire-time 0 "
-      NotifyTimeoutSeconds s ->
-        mconcat
-          [ " --expire-time ",
-            showt (fromIntegral @_ @Integer s * 1_000)
-          ]
+      sendNote (DBus client) = MonadDBus.notifyDBus client note
+      sendNote NotifySend = MonadNotifySend.notifyNotifySend note
+      sendNote AppleScript = MonadAppleScript.notifyAppleScript note